making strings signed char*

+++ lib/buffer/buffer_dump.c
@@ -4,9 +4,9 @@
 
 #ifdef DEBUG
 
-extern int term_read(int fd, char *buf, unsigned int len);
-extern int buffer_dummyreadmmap();
-extern int stralloc_write();
+extern ssize_t term_read(int fd, char *buf, unsigned int len);
+extern ssize_t buffer_dummyreadmmap();
+extern ssize_t stralloc_write();
 
 void buffer_dump(buffer *out, buffer *b) 
 {
